- Bugfix: DBChampSpec table is now included in the spell update - Bugfix: Various casts and usages of int spellID instead of ushort. - Bugfix: Fixed an issue where players couldn't see any other players or NPCs after jumping away from a house without exiting it properly, (teleports and such)
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/EtaewMajor@1160 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/effects/GameEffectList.cs
@@ -164,7 +164,8 @@ namespace DOL.GS.Effects
 			foreach (PlayerXEffect eff in effs)
 			{
 				bool good = true;
-				Spell spell = SkillBase.GetSpellByID(eff.Var1);
+				#warning There is a danger here of someone having a spell entry > ushort.Max here, and it being cast to ushort
+				Spell spell = SkillBase.GetSpellByID((ushort)eff.Var1);
 				if (spell == null)
 					good = false;
